Move ZookeeperClient to the kafka.zookeeper package
