+++ core/src/test/scala/unit/kafka/controller/ZookeeperClientTest.scala
@@ -23,6 +23,7 @@ import java.util.concurrent.{CountDownLatch, TimeUnit}
 import javax.security.auth.login.Configuration
 
 import kafka.zk.ZooKeeperTestHarness
+import kafka.zookeeper._
 import org.apache.kafka.common.security.JaasUtils
 import org.apache.zookeeper.KeeperException.Code
 import org.apache.zookeeper.{CreateMode, ZooDefs}
